pyproject.toml
[build-system]
requires = [
    "setuptools>=60.0",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "agentzero-cli"
version = "0.2.0"
description = "AI coding agent with security interceptor - TUI/CLI with Local LLM, OpenRouter support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Wojciech Wiesner", email = "wojciech@theones.io"},
]
keywords = [
    "ai-agent",
    "coding-assistant",
    "cli",
    "tui",
    "llm",
    "terminal",
    "security",
    "human-in-the-loop",
    "textual",
    "openrouter",
    "ollama",
    "lm-studio",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Terminals",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "textual>=0.40.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "build",
    "twine",
]

[project.urls]
Homepage = "https://github.com/vizi2000/agentzero-cli"
Repository = "https://github.com/vizi2000/agentzero-cli"
Documentation = "https://github.com/vizi2000/agentzero-cli/tree/main/docs"
Issues = "https://github.com/vizi2000/agentzero-cli/issues"
Changelog = "https://github.com/vizi2000/agentzero-cli/blob/main/CHANGELOG.md"

[project.scripts]
a0 = "agentzero_cli.main:main"
a0tui = "agentzero_cli.main:main"
a0cli = "agentzero_cli.cli_entry:main"
agentzero = "agentzero_cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentzero_cli*"]

[tool.setuptools.package-data]
agentzero_cli = ["*.yaml", "*.example"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
